Moves audits and flattens hierarchy
