+++ src/audits/manifest-short-name-length.js
@@ -17,7 +17,7 @@
 
 'use strict';
 
-const Audit = require('../audit');
+const Audit = require('./audit');
 
 class ManifestShortNameLength extends Audit {
   /**
