+++ src/audits/manifest-start-url.js
@@ -17,7 +17,7 @@
 
 'use strict';
 
-const Audit = require('../audit');
+const Audit = require('./audit');
 
 class ManifestStartUrl extends Audit {
   /**
