+++ src/audits/service-worker.js
@@ -17,7 +17,7 @@
 
 'use strict';
 
-const Audit = require('../audit');
+const Audit = require('./audit');
 
 class ServiceWorker extends Audit {
   /**
